remove note about being included in libmusicbrainz
The basic discid code was in libmusicbrainz3,
but it was removed in libmusicbrainz4 and lmb3 is deprecated since 2011.

Additionally, libmusicbrainz3 obviously didn't include anything on
reading ISRCs from disc.

I also included the direct link to the musicbrainz ticket component to
make it easier to find the correct place to report bugs.