feature names to lowercase; room for 32 features
It is a lot easier to get the feature names right
when you don't have to think about casing.

We will probably never reach 16 features,
but better safe than sorry.
Having an array of 32 pointers doesn't hurt.
Additionally, that is the same room we would have
when using bit masks in an int (mostly 4 byte).