autotools: libtool version-info -> version-number
The calculation of current:revision:age for libtool version-info
is error-prone.
We now use MAJOR.MINOR.PATCH as defined in http://semver.org,
but we start with that already at 0.0.0, not with version 1.

The result is the same. Libtool takes the -version-number parameter
instead.

Additionally we only have two places were we define version numbers:
The main cmake and autotools files.
We have to define the version twice in configure.ac
due to AC_INIT limitations/warnings.

The corresponding ticket is
http://tickets.musicbrainz.org/browse/LIB-32