add DISCID_VERSION_NUM for easy calculation
This is basically one decimal number created using two digits for each
part of the version.
This breaks when patch version > 99 (unlikely) and has to be adapted when
the major version changes (> 0).

MAJOR is not used now, since we can't have leading zeroes.
This would make it an octal number.