remove read_sparse contradiction in the docs
Since DISCID_FEATURE_READ == 0, you automatically specify this feature.
So we should explicitely say, that read_sparse always reads the TOC.

For the same reason we can't and we don't have to test
if the feature is requested in the code.
`features & 0` is always `0`.

Note that `has_feature(DISCID_FEATURE_READ)` doesn't work with bit masks,
but is implemented with switch.