add ntddcdrm.h for MSVC
MS Visual Studio doesn't seem to include that header by itself.
The header is included in a separate Windows Driver Development Kit,
but that one is difficult to install and possibly not redistributable.

The ntddcdrm.h is taken from the dosbox repository.
MinGW and cygwin do include the header. MinGW only works with the MinGW
header.