MSC: use _CRT_SECURE_NO_WARNINGS
While warning for buffer overflows etc. is good in general,
using unportable string functions is not worth it.
The current usage is no actual problem.

However, I put the define in every file, so we know which files are
possibly affected.

_CRT_SECURE_CPP_OVERLOAD_STANDARD_NAMES (_COUNT) didn't help at all.
The NO_WARNINGS define was called _CRT_SECURE_NO_DEPRECATE
in earlier versions of Visual Studio.