don't put mb_disc_read_unportable in unix.c
We explicitely call it in the platform files as mb_disc_unix_read().
This removes some "black magic" since all other unportable functions
are also implemented in the platform files.

This is also a prerequisite to accept numbers as drive names for Mac OS X,
since we need to intercept the device name before opening it.