implement CRC check for raw ISRCs
I implemented this from scratch.
This is probably not the most efficient implementation,
but the check doesn't even have a measurable impact.
We are reading data from disc and only check the CRC when an ISRC is
found, which is ca. every 100th frame.

The focus is a clear implementation without storing (CRC-8) tables in the code.

CRC mismatches are only printed to stderr for now.