raw isrcs: don't return main channel data
This doesn't have an impact on the performance,
since the data is still read from disc, just not returned by the
command.
However, there seems to be a limit of data returned with one command.
My tests failed after reaching something over 683 KB.

This enables us to read data for up to 71 ISRCs,
but that would also take ca. 5 minutes per disc.

Setting this to 2, to tolerate one failed CRC check.