fix warning for empty m4 directory
This removes the warning about a missing m4 directory,
which is empty because we don't have any macros
in the repository.
The warning broke the git build in Ubuntu.

Now thert is a message that we should "consider"
adding the m4 commands we just removed.
Hopefully this is not a breaking warning for Ubuntu.